// src/types/r3eTypes.ts
export interface ITireData<T> {
	FrontLeft: T;
	FrontRight: T;
	RearLeft: T;
	RearRight: T;
}
// ...
export interface ITireTemp {
	CurrentTemp: {
		Left: number;
		Center: number;
		Right: number;
	};
	OptimalTemp: number;
	ColdTemp: number;
	HotTemp: number;
}

export interface IBrakeTemp {
	CurrentTemp: number;
	OptimalTemp: number;
	ColdTemp: number;
	HotTemp: number;
}